change the class name HttpClientImpl and its method from Future to Stream
