+++ lib/core/http_client/http_client_requests.dart
@@ -1,0 +1,10 @@
+abstract class HttpClientRequests {
+  Stream<HttpResponse> get(String url);
+}
+
+class HttpResponse {
+  final int? statusCode;
+  final dynamic data;
+
+  HttpResponse({required this.statusCode, required this.data});
+}
