+++ lib/modules/curiosity/data/repositories/curiosity_repository.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:injectable/injectable.dart';
-import 'package:taplingo_matrix/core/http_client/http_client.dart';
+import 'package:taplingo_matrix/core/http_client/http_client_requests.dart';
 import 'package:taplingo_matrix/core/utils/keys/nasa_keys.dart';
 import 'package:taplingo_matrix/modules/curiosity/data/datasources/endpoints/curiosity_endpoints.dart';
 import 'package:taplingo_matrix/modules/curiosity/data/models/curiosity_model.dart';
@@ -9,14 +9,15 @@ import 'package:taplingo_matrix/modules/curiosity/domain/ports/output/get_mars_i
 
 @singleton
 class CuriosityRepository implements GetMarsImagesOutputPort {
-  final HttpClient client;
+  final HttpClientRequests client;
 
   CuriosityRepository(this.client);
 
   @override
   Stream<List<CuriosityModel>> getMarsImages() async* {
-    HttpResponse response = await client
+    Stream<HttpResponse> responses = await client
         .get((CuriosityEndpoints.getMarsImagesURL(NasaKeys.apiKey)).toString());
+    HttpResponse response = await responses.first;
     if (response.statusCode == 200) {
       Map<String, dynamic> data =
           Map<String, dynamic>.from(json.decode(response.data));
